% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callConcourse.R
\name{callConcourse}
\alias{callConcourse}
\title{Call Concourse}
\usage{
callConcourse(url, env_var_name = "CONCOURSE_KEY", verbose = FALSE)
}
\arguments{
\item{url}{The url constructed for the API call}

\item{env_var_name}{The name of the environment variable of the API key}

\item{verbose}{Enable verbose mode}
}
\value{
content of the API call as a list
}
\description{
Call Concourse
}
