% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChangeLog.R
\name{getChangeLog}
\alias{getChangeLog}
\title{Get Concourse Change Log Info}
\usage{
getChangeLog(id, type = "external_id", server = "sandbox", ...)
}
\arguments{
\item{id}{An internal or external id for the course}

\item{type}{Options are "external_id" or "course_id"}

\item{server}{Options are "sandbox", "production", or a provided URL}

\item{...}{Optional info for callConcourse}
}
\description{
Get Concourse Change Log Info
}
