% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCourseIDs.R
\name{getCourseIDs}
\alias{getCourseIDs}
\title{Get Concourse Course IDs}
\usage{
getCourseIDs(
  keyword = NULL,
  keyword_mode = NULL,
  title = NULL,
  prefix = NULL,
  number = NULL,
  campus_id = NULL,
  school_id = NULL,
  department_id = NULL,
  instructor = NULL,
  session = NULL,
  year = NULL,
  timeframe = NULL,
  delivery = NULL,
  credits = NULL,
  template = NULL,
  audit_status = NULL,
  server = "sandbox",
  ...
)
}
\arguments{
\item{keyword}{String of keyword}

\item{keyword_mode}{Options are "any" or "all"}

\item{title}{String of title}

\item{prefix}{String of course prefix}

\item{number}{String of course number/code}

\item{campus_id}{String of campus ID}

\item{school_id}{String of college/school ID}

\item{department_id}{String of department ID}

\item{instructor}{String of instructor name}

\item{session}{String of session}

\item{year}{String of year}

\item{timeframe}{Options are "past", "past_current", "current", "current_future", or "future"}

\item{delivery}{String of delivery}

\item{credits}{String of the number of course credits}

\item{template}{Options are "templates" or "non"}

\item{audit_status}{One of "0", "1", "2", "3", or "4".  1 means In Progress, 2 means Reviewed, 3 means Modified Since Review, and 4 means Submitted For Review.}

\item{server}{Options are "sandbox", "production", or a provided URL}

\item{...}{Optional info for callConcourse}
}
\description{
Get Concourse Course IDs
}
