% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSyllabusAccess.R
\name{getSyllabusAccess}
\alias{getSyllabusAccess}
\title{Get Concourse Syllabus Access Data}
\usage{
getSyllabusAccess(id, type = "external_id", server = "sandbox", ...)
}
\arguments{
\item{id}{An internal or external id for the course}

\item{type}{Options are "external_id" or "course_id"}

\item{server}{Options are "sandbox", "production", or a provided URL}

\item{...}{Optional info for callConcourse}
}
\description{
Get Concourse Syllabus Access Data
}
